organize legacy minter filter tests
