+++ test/minter-filter/v0-v1/enumeration/MinterFilterEnumeration.common.ts
@@ -1,7 +1,7 @@
 import { expectRevert } from "@openzeppelin/test-helpers";
 import { expect } from "chai";
 import { loadFixture } from "@nomicfoundation/hardhat-network-helpers";
-import { safeAddProject, T_Config } from "../../util/common";
+import { safeAddProject, T_Config } from "../../../util/common";
 
 /**
  * These tests are intended to check common Enumeration behaviors of
