+++ test/minter-filter/v0-v1/enumeration/MinterFilterV1Enumeration.test.ts
@@ -1,4 +1,3 @@
-import { loadFixture } from "@nomicfoundation/hardhat-network-helpers";
 import {
   T_Config,
   getAccounts,
@@ -6,9 +5,10 @@ import {
   deployAndGet,
   deployCoreWithMinterFilter,
   safeAddProject,
-} from "../../util/common";
-
-import { MinterFilterPermissions_Common } from "./MinterFilterPermissions.common";
+} from "../../../util/common";
+import { expectRevert } from "@openzeppelin/test-helpers";
+import { loadFixture } from "@nomicfoundation/hardhat-network-helpers";
+import { MinterFilterEnumeration_Common } from "./MinterFilterEnumeration.common";
 
 const runForEach = [
   {
@@ -24,14 +24,14 @@ const runForEach = [
     minter: "MinterSetPriceERC20V2",
   },
 ];
-
 runForEach.forEach((params) => {
-  describe(`${params.minterFilter} Permissions w/${params.core} core`, async function () {
+  describe(`${params.minterFilter} Enumeration w/${params.core} core`, async function () {
     async function _beforeEach() {
       let config: T_Config = {
         accounts: await getAccounts(),
       };
       config = await assignDefaultConstants(config);
+
       // deploy and configure minter filter and minter
       ({
         genArt721Core: config.genArt721Core,
@@ -41,7 +41,6 @@ runForEach.forEach((params) => {
         params.core,
         params.minterFilter
       ));
-
       config.minter = await deployAndGet(config, params.minter, [
         config.genArt721Core.address,
         config.minterFilter.address,
@@ -53,15 +52,28 @@ runForEach.forEach((params) => {
         config.accounts.deployer,
         config.accounts.artist.address
       );
-
-      await config.genArt721Core
-        .connect(config.accounts.artist)
-        .updateProjectMaxInvocations(config.projectZero, config.maxInvocations);
       return config;
     }
 
     describe("common tests", async function () {
-      await MinterFilterPermissions_Common(_beforeEach);
+      await MinterFilterEnumeration_Common(_beforeEach);
+    });
+
+    describe("test specific to V1", async function () {
+      it("doesn't allow removal of unapproved minters", async function () {
+        const config = await loadFixture(_beforeEach);
+        if (params.minterFilter !== "MinterFilterV1") {
+          console.log("skipping test for non-V1 minter filter");
+          return;
+        }
+        // reverts when attempting to remove minter being used
+        await expectRevert(
+          config.minterFilter
+            .connect(config.accounts.deployer)
+            .removeApprovedMinter(config.minter.address),
+          "Only approved minters"
+        );
+      });
     });
   });
 });
